backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email]
langchain
langchain-community
langchain-text-splitters
langchain-ollama
ollama
qdrant-client
faster-whisper
python-dotenv
pypdf
reportlab
fpdf2
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
langgraph
sqlalchemy
alembic
